fix(cache): watch errors must call done handler
The type of `watchObject` was incorrect and has been updated to match
the actual request body.

Using this info it was clear that 'ERROR' events were not being handled
correctly. When the watch receives an error it is not always an http
status code, because the status code can only be sent when the stream is
starting. This means that `410` resourceVersion out of date errors could
only be handled if they were detected before the watch stream started
leaving watches running on channels that would never receive more events
and not notifying `ListWatch` consumers of the error.